Refactor large frontend and backend components following SOLID principles
- Broke down `BatchJobTable.tsx` into modular sub-components (Headers, Rows, Pagination, States) and extracted shared utility configurations.
- Decomposed `SystemHealth.tsx` into individual metric widgets and shared helper modules.
- Extracted bulky user deletion data cleanup logic from `IdentityService.cs` into a dedicated `UserCleanupUtility` to adhere to the Single Responsibility Principle.
- Re-tested all modified applications to ensure no functional regressions were introduced.
